menambahkan validasi untuk hapus data produk dan master

+++ app/Http/Controllers/AdminToko/MasterJenisBarangController.php
@@ -98,7 +98,16 @@ class MasterJenisBarangController extends Controller
     {
         $item = Type::findOrFail($id);
 
+        if (
+            $item->relasiService()->exists()
+        ) {
+            toast('Data Jenis Barang yang memiliki riwayat transaksi tidak bisa dihapus.', 'error');
+            return redirect()->back();
+        }
+
         $item->delete();
+
+        toast('Data Jenis Barang berhasil dihapus.', 'success');
 
         return redirect()->route('admin-master-jenis-barang.index');
     }
